[FE] Refactor: PlayControllerButton/Sub,MainMenu index.tsx에서 컴포넌트 정의
- 기존의 index.ts -> index.tsx로 변경 후 컴포넌트 정의 및 default export
- global css a 태그 속성 삭제
