+++ client/assets/global.css
@@ -19,10 +19,6 @@ a {
     text-decoration: none;
 }
 
-a:hover {
-    text-decoration: underline;
-}
-
 button {
     border: none;
     background-color: inherit;
