+++ client/components/molecules/MainMenu/MainMenu.stories.tsx
@@ -1,9 +1,9 @@
-import React from "react";
-import MainMenu from "./MainMenu"
+import React from 'react';
+import MainMenu from './index';
 
 export default {
-  title: "MainMenu",
-  component: MainMenu,
+    title: 'MainMenu',
+    component: MainMenu,
 };
 
-export const Default = () =>  <MainMenu />;
+export const Default = () => <MainMenu />;
