--- client/components/molecules/MainMenu/index.ts
@@ -1,1 +1,0 @@
-export { default } from './MainMenu';
