--- client/components/molecules/PlayControllerButtons/index.ts
@@ -1,1 +1,0 @@
-export { default } from './PlayControllerButtons';
