--- client/components/molecules/SubMenu/index.ts
@@ -1,1 +1,0 @@
-export { default } from './SubMenu';
